add float to double fix

+++ DataAccess/CaseServiceRepository.cs
@@ -149,14 +149,14 @@ namespace DataAccess
                             CaseService caseService = new CaseService();
                             caseService.CaseID = (int)reader["caseID"];
                             caseService.ServiceID = (int)reader["serviceID"];
-                            caseService.Hours = (float)reader["hours"];
-                            caseService.Kilometres = (float)reader["kilometres"];
-                            caseService.EstimatedHours = (float)reader["estimatedHours"];
+                            caseService.Hours = (double)reader["hours"];
+                            caseService.Kilometres = (double)reader["kilometres"];
+                            caseService.EstimatedHours = (double)reader["estimatedHours"];
 
                             Service service = new Service();
                             service.ID = (int)reader["ID"];
                             service.Name = (string)reader["name"];
-                            service.Price = (float)reader["price"];
+                            service.Price = (double)reader["price"];
                             service.IsHourly = (bool)reader["isHourly"];
 
                             servicesByCaseService.Add(caseService, service);
@@ -213,9 +213,9 @@ namespace DataAccess
         {
             caseService.CaseID = (int)reader[0];
             caseService.ServiceID = (int)reader[1];
-            caseService.Hours = (float)reader[2];
-            caseService.Kilometres = (float)reader[3];
-            caseService.EstimatedHours = (float)reader[3];
+            caseService.Hours = (double)reader[2];
+            caseService.Kilometres = (double)reader[3];
+            caseService.EstimatedHours = (double)reader[3];
         }
         private static List<CaseService> MapCollection(SqlCommand command)
         {
